# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: manualbox API
  version: "0.1.0-m0"
  description: |
    The manualbox HTTP API.

    This document is the written contract for the API. The TypeScript client types
    in `web/src/api/types.ts` are maintained by hand against it — with ten
    endpoints, a generator cost more than it saved (and pulled in a dependency
    subtree with its own advisories). Revisit generation when M1 multiplies the
    surface.

    **Authentication.** Every endpoint except `/health`, `/setup`, and
    `/auth/login` requires a session. The session cookie is `manualbox_session`
    (HttpOnly, SameSite=Lax, Secure behind TLS). Scripts and the CLI may instead
    send `Authorization: Bearer <token>`.

    **CSRF.** State-changing requests carrying an `Origin` header that matches
    neither the request host nor the configured base URL are rejected with 403
    `cross_origin_blocked`. Requests with no `Origin` (curl, native clients) are
    allowed, since there is no ambient cookie to abuse.

  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:7745/api/v1
    description: Local instance
...
paths:
  /search:
    get:
      tags: [search]
      summary: Which manual says X, and where
      description: |
        Full-text search over every converted manual in the household. A hit names
        the document, the device, the page and the language, and carries enough text
        to recognise — "page 47 of something" does not solve the problem this
        endpoint exists for.

        A GET with `?q=`, so a search is a URL: linkable, bookmarkable, and in the
        browser's history where a user expects it.

        **`q` is text, never a query language.** SQLite FTS5 has an expression syntax
        — bare `AND`, `OR`, `NOT`, `NEAR`, colons, prefix stars — and passing a
        user's words to it would make an ordinary query a syntax error and silently
        reinterpret another. Each word is quoted as a phrase and the phrases are
        ANDed, so two words mean a block containing both.

        **Two modes, and the response says which one answered.** The index is a
        trigram index, chosen because the corpus is not English: Chinese, Japanese
        and Thai do not put spaces between words, and a word-boundary tokeniser
        finds nothing at all inside them. Its cost is that no token shorter than
        three characters exists in it, so a query where any word is shorter than
        that is answered by scanning instead — `mode` is then `substring`, and
        `bm25` and `score` are 0 on every hit because there is no index term to
        weigh. See `internal/db/migrations/00006_block_search.sql` for the
        measurements.

        A query matching nothing returns `indexed`, which is how many blocks exist
        to search at all: "no manual says that" and "nothing has been converted yet"
        are otherwise the same empty list.
      parameters:
        - name: q
          in: query
          required: true
          description: What to look for, as typed. Whitespace-only is refused rather than answered.
          schema: { type: string, examples: [Saugkraft] }
        - name: documentId
          in: query
          required: false
          description: |
            Narrow to one manual, which is what a reader already inside a document
            asks. Omitted, the search spans every document, because a household
            looking for the descaling interval does not know which manual to open.

            An unknown id is a search of nothing rather than a 404: this parameter
            scopes a search, and answering 404 would turn it into a way to test
            whether an id exists.
          schema: { type: string, examples: [doc_01JQ8ZK3M4N5P6R7S8T9V0W1X2] }
        - name: limit
          in: query
          required: false
          description: 1 to 100. Defaults to 25.
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: The hits
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResults" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
...
  /setup:
    get:
      tags: [setup]
      summary: Whether first-run setup is still needed
      security: []
      responses:
        "200":
          description: Setup status
          content:
            application/json:
              schema:
                type: object
                required: [needsSetup]
                properties:
                  needsSetup: { type: boolean }
    post:
      tags: [setup]
      summary: Create the first administrator
      description: |
        Only succeeds while the instance has no users; afterwards it returns 409.
        On success the new administrator is signed in and a session cookie is set,
        so the first-run flow does not ask for the password twice.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 10 }
                displayName: { type: string, description: Defaults to the email address }
      responses:
        "201":
          description: Administrator created and signed in
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/Error" }
        "409":
          description: Setup already completed (`setup_complete`)
          $ref: "#/components/responses/Error"
        "422":
          description: Weak password (`weak_password`) or invalid email (`invalid_email`)
          $ref: "#/components/responses/Error"
  /auth/login:
    post:
      tags: [auth]
      summary: Sign in
      description: |
        Returns an identical error whether the address is unknown or the password
        is wrong, and takes the same time in both cases, so registered addresses
        cannot be enumerated.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Signed in; sets the session cookie
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "401":
          description: Wrong credentials (`invalid_credentials`)
          $ref: "#/components/responses/Error"
...
  /auth/me:
    get:
      tags: [auth]
      summary: The current user and session
      responses:
        "200":
          description: Current identity
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  session: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/password:
    post:
      tags: [auth]
      summary: Change the password
      description: Revokes every session, including the caller's.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 10 }
      responses:
        "204": { description: Changed; all sessions revoked }
        "401":
          description: Current password is wrong (`invalid_credentials`)
          $ref: "#/components/responses/Error"
        "422":
          description: New password too weak (`weak_password`)
          $ref: "#/components/responses/Error"
  /instance:
    get:
      tags: [system]
      summary: What this deployment can currently do
      description: |
        Lets the UI hide features with no configured provider rather than offering
        a button that always fails. Never includes provider API keys — only
        whether one is set.
      responses:
        "200":
          description: Instance capabilities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Instance" }
        "401": { $ref: "#/components/responses/Error" }
  /jobs:
    get:
      tags: [jobs]
      summary: Recent background jobs, newest first
      parameters:
        - name: state
          in: query
          schema: { $ref: "#/components/schemas/JobState" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: object
                required: [jobs]
                properties:
                  jobs:
                    type: array
                    items: { $ref: "#/components/schemas/Job" }
        "400":
          description: Bad `state` (`invalid_state`) or `limit` (`invalid_limit`)
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Error" }
  /jobs/events:
    get:
      tags: [jobs]
      summary: Live job progress (Server-Sent Events)
      description: |
        Opens an SSE stream of `job` events. On connect the current queued and
        running jobs are replayed, so a client joining mid-conversion sees state
        immediately and a reconnect re-synchronises without bookkeeping.

        Comment-only keepalives are sent every 25 seconds so idle proxies do not
        close a quiet stream.
      responses:
        "200":
          description: An event stream
          content:
            text/event-stream:
              schema: { type: string }
              example: |
                event: job
                data: {"jobId":"job_01JQ…","kind":"translate","state":"running","progress":0.67,"note":"translating section 2 of 3","at":"2026-07-25T11:32:19.699Z"}
        "401": { $ref: "#/components/responses/Error" }
  /jobs/{jobID}:
    parameters:
      - name: jobID
        in: path
        required: true
        schema: { type: string, examples: [job_01JQ8ZK3M4N5P6R7S8T9V0W1X2] }
    get:
      tags: [jobs]
      summary: One job
      responses:
        "200":
          description: The job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /jobs/{jobID}/cancel:
    parameters:
      - name: jobID
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [jobs]
      summary: Cancel a queued or running job
      description: |
        A running handler is not killed; it observes cancellation at its next
        progress report. Interrupting a handler mid-write could leave a document
        half-converted, so cancellation is cooperative.
      responses:
        "204": { description: Cancelled }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: Already finished (`not_cancellable`)
          $ref: "#/components/responses/Error"
  /documents/{documentID}/gate:
    parameters:
      - name: documentID
        in: path
        required: true
        schema: { type: string, examples: [doc_01JQ8ZK3M4N5P6R7S8T9V0W1X2] }
    get:
      tags: [documents]
      summary: The pre-flight question, before anything is spent
      description: |
        What manualbox is holding, what it would process, and what that would
        cost. Built entirely from stored probe results, so it survives a restart,
        costs nothing to render, and never re-reads the document.

        The rest of the document and device surface is not written up here yet;
        this path is documented because the gate is the screen a user decides on
        and its shape is what the SPA reads.
      responses:
        "200":
          description: The gate
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Gate" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /documents/{documentID}/approve:
    parameters:
      - name: documentID
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [documents]
      summary: Authorise the conversion the gate described
      description: |
        The moment the user agrees to spend, and the opposite of `decline`. The
        document moves to `converting` and a `doc.convert` job is queued; progress
        arrives on the existing job event stream.

        **There is no request body, and there must not be one.** The scope
        converted is the household's configured reading languages, which is what the
        gate showed the user. Accepting a language list here would let the thing
        approved differ from the thing shown, which is the one promise the funnel
        makes.

        Refused with `invalid` when there is nothing honest to convert: the document
        has not been read yet, it is password-protected, it has no text layer and so
        needs OCR, or none of its languages are ones this household reads. A refusal
        does not move the document.
      responses:
        "202":
          description: Queued
          content:
            application/json:
              schema:
                type: object
                required: [document]
                properties:
                  document: { type: object, description: The document, now in `converting`. }
                  jobId: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /documents/{documentID}/figures/{sha256}:
    parameters:
      - name: documentID
        in: path
        required: true
        schema: { type: string }
      - name: sha256
        in: path
        required: true
        description: The figure's digest, as `Figure.sha256` reports it.
        schema: { type: string, pattern: "^[0-9a-f]{64}$" }
    get:
      tags: [documents]
      summary: One rendered figure's PNG
      description: |
        Addressed by digest because the content is the name: the ETag is exact and
        the bytes are immutable for ever.

        The digest is checked against this document's own figures rather than handed
        to the blob store directly. The store holds every original anyone has
        uploaded, so a route that opened any digest a caller named would serve
        another household's manual to whoever could guess a hash.
      responses:
        "200":
          description: The PNG
          content:
            image/png:
              schema: { type: string, format: binary }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: manualbox_session
    bearerToken:
      type: http
      scheme: bearer
      description: The same session token, for scripts and the CLI.
  responses:
    Error:
      description: A failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      description: The single error shape every failing endpoint returns.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable; safe to branch on.
              examples: [invalid_credentials, not_found, weak_password]
            message:
              type: string
              description: Human-readable; not stable, do not branch on it.
    User:
      type: object
      required: [id, email, displayName, role, createdAt]
      properties:
        id: { type: string, examples: [usr_01JQ8ZK3M4N5P6R7S8T9V0W1X2] }
        email: { type: string, format: email }
        displayName: { type: string }
        role: { type: string, enum: [admin, member, viewer] }
        createdAt: { type: string, format: date-time }
        lastLoginAt: { type: string, format: date-time }
    Session:
      type: object
      required: [id, userId, createdAt, lastSeenAt, expiresAt]
      properties:
        id: { type: string }
        userId: { type: string }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        expiresAt:
          type: string
          format: date-time
          description: Slides forward on use, so day-to-day use never forces a re-login.
    JobState:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    Job:
      type: object
      required: [id, kind, state, progress, attempts, maxAttempts, createdAt, updatedAt]
      properties:
        id: { type: string }
        kind:
          type: string
          description: What the job does.
          examples: [doc.probe, doc.convert]
        state: { $ref: "#/components/schemas/JobState" }
        priority: { type: integer, description: Higher runs first. }
        progress: { type: number, minimum: 0, maximum: 1 }
        progressNote: { type: string, examples: ["translating section 2 of 3"] }
        attempts: { type: integer }
        maxAttempts: { type: integer }
        lastError: { type: string }
        tokensIn: { type: integer, description: Provider input tokens consumed. }
        tokensOut: { type: integer }
        costMicros:
          type: integer
          description: Provider cost in millionths of a currency unit, to keep money out of floats.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
...
    ProviderInfo:
      type: object
      description: A configured adapter. Never carries the API key itself.
      required: [kind, model, enabled, hasAPIKey]
      properties:
        kind: { type: string, examples: [poppler, tesseract, claude, ollama, deepl] }
        model: { type: string }
        enabled: { type: boolean, description: False means the feature is switched off. }
        hasAPIKey: { type: boolean }
    ExternalTool:
      type: object
      required: [available, version, purpose, install]
      properties:
        available: { type: boolean }
        version: { type: string }
        purpose: { type: string }
        install: { type: string, description: Platform-specific install hint. }
    Instance:
      type: object
      required: [version, languages, primaryLanguage, capabilities, providers, externalTools]
      properties:
        version: { type: string }
        languages:
          type: array
          items: { type: string }
          description: BCP-47 reading languages, most preferred first.
        primaryLanguage: { type: string }
        capabilities:
          type: object
          properties:
            convert: { type: boolean }
            ocr: { type: boolean }
            translate: { type: boolean }
            extract: { type: boolean }
        providers:
          type: object
          additionalProperties: { $ref: "#/components/schemas/ProviderInfo" }
        externalTools:
          type: object
          additionalProperties: { $ref: "#/components/schemas/ExternalTool" }
...
    Gate:
      type: object
      description: |
        The pre-flight answer for one document.

        The language map is read from the stored regions where there are regions,
        and from the per-page runs otherwise. On a parallel-columns manual a page
        holds three languages, so no per-page answer about it can be right; on a
        sequential manual the two sources agree.
      required:
        [documentId, deviceId, kind, state, probed, pages, encrypted, hasTextLayer,
         medianChars, chars, household, inScope, other, scopePages, scopeFraction,
         scopeChars, scopeCharFraction, conflicts, unlabelledPages, requiresApproval,
         maxPagesAuto, cost, summary]
      properties:
        documentId: { type: string }
        deviceId: { type: string }
        filename: { type: string }
        kind: { type: string, enum: [manual, receipt, warranty, photo, other] }
        state: { type: string, enum: [uploaded, probing, awaiting_scope, declined, converting, ready, failed] }
        probed: { type: boolean }
        pages: { type: integer }
        encrypted: { type: boolean }
        hasTextLayer: { type: boolean, description: False means this is a scan and reading it needs OCR. }
        medianChars: { type: integer }
        chars:
          type: integer
          description: |
            The document's named text: the characters of every language something
            could name, and the denominator of every `share`. Text nothing could
            name is excluded, so a failed signal cannot silently shrink a share.
        household:
          type: array
          items: { type: string }
          description: The configured reading languages, echoed back so the UI can explain scope.
        inScope:
          type: array
          items: { $ref: "#/components/schemas/GateLanguage" }
        other:
          type: array
          items: { $ref: "#/components/schemas/GateLanguage" }
          description: |
            Present but not read by this household. Listed, never discarded: the
            original is kept whole, so importing one later is a button rather than
            a re-upload.
        scopePages:
          type: integer
          description: |
            Distinct pages carrying an in-scope language, not a sum over
            languages — on a columns manual a sum reports 133 pages of a 68-page
            document.
        scopeFraction: { type: number, description: "`scopePages` over `pages`." }
        scopeChars: { type: integer }
        scopeCharFraction:
          type: number
          description: |
            `scopeChars` over `chars`. The honest measure of how much of a
            document a household reads: on the measured columns manual German is
            38% by pages and 20% by characters.
        conflicts:
          type: integer
          description: |
            How many runs the signals disagreed about — the document's own contents
            table against its pages. A region's disagreement (a column's alphabet
            against the page's printed tab) is a different thing and reaches the
            user through `conflict` on the language itself.
        unlabelledPages:
          type: integer
          description: |
            Content pages carrying text that no signal could name. Front matter and
            a back cover are excluded: they legitimately belong to no section.
        requiresApproval: { type: boolean }
        maxPagesAuto: { type: integer }
        cost: { $ref: "#/components/schemas/CostEstimate" }
        summary: { type: string, description: One line, characters-led, for a person to read. }
...
    SearchResults:
      type: object
      description: The hits, plus what was actually asked and how it was answered.
      required: [query, mode, limit, truncated, hits]
      properties:
        query: { type: string, description: The text as typed, echoed so a client need not keep its own copy in step. }
        mode:
          type: string
          enum: [index, substring]
          description: |
            Which path answered. `index` is the FTS5 trigram index with bm25 ranking.
            `substring` is the scan that covers a query the index cannot represent —
            one where some word is shorter than three characters — and it does not
            rank.
        limit: { type: integer, description: The limit applied, after clamping to 100. }
        truncated:
          type: boolean
          description: |
            The limit cut the list off, which is the difference between "these are
            the hits" and "these are the first hits".
        hits:
          type: array
          items: { $ref: "#/components/schemas/SearchHit" }
        indexed:
          type: integer
          description: |
            How many blocks exist to search. Present **only** when nothing matched,
            because that is when "no manual says that" and "nothing has been
            converted yet" are otherwise indistinguishable.
    SearchHit:
      type: object
      description: |
        One match. `page`, `regionX0` and `index` are the block's natural key — the
        same citation `Block` carries — so a hit can be deep-linked to the exact
        paragraph and still points there after a re-conversion.
      required: [documentId, deviceId, deviceName, state, page, regionX0, index, kind, snippet, chars, bm25, score]
      properties:
        documentId: { type: string }
        filename: { type: string, description: What the uploaded file was called. Half of "which manual". }
        deviceId: { type: string }
        deviceName: { type: string, description: The other half. A household recognises the device, not the id. }
        state:
          type: string
          enum: [uploaded, probing, awaiting_scope, declined, converting, ready, failed]
          description: The document's pipeline state, so a hit from a manual that is mid-re-conversion is visible as such.
        page: { type: integer, description: 1-based PDF page, the number the paper means. }
        regionX0: { type: integer }
        index: { type: integer }
        kind:
          type: string
          enum: [heading, paragraph, list-item, table, figure]
        level: { type: integer }
        lang: { type: string }
        name: { type: string, description: "`lang` for a person to read: Japanese, not ja." }
        snippet:
          type: string
          description: |
            About 64 characters of the block around the match — enough to recognise.
            `chars` is the whole block's length, so a client can tell a snippet from
            a complete block and fetch the rest from
            `/documents/{documentID}/conversion`.
        chars: { type: integer, description: The block's full length in runes, not bytes. }
        bm25:
          type: number
          description: |
            FTS5's own relevance, negative and lower-is-better. 0 in `substring`
            mode, where there is no index term to weigh.
        score:
          type: number
          description: |
            What the results are ordered by: `bm25` minus 1.0 for a heading. The
            heading bonus is a judgement — a heading names a section and so answers
            "where does it say this" better than a passing mention — and both numbers
            are reported so it can be argued with rather than merely trusted.
security:
  - sessionCookie: []
  - bearerToken: []